+++ packages/i18n-core/src/TranslationManager.ts
@@ -1,5 +1,4 @@
-import { escapeHTMLTags } from "./utils/escapeHTMLTags.util";
-import { interpolate } from "./utils/interpolate.util";
+import type { ILooseObject } from "./types";
 
 export default class TranslationManager<
   TLanguage extends string,
@@ -28,18 +27,42 @@ export default class TranslationManager<
           currentObject = currentObject[part as keyof typeof currentObject];
         }
         if (params.interpolation) {
-          return interpolate(
+          return this.interpolate(
             currentObject,
             params.interpolation,
             params.escapeHTML,
           );
         }
         if (typeof params.escapeHTML === "boolean" && params.escapeHTML)
-          return escapeHTMLTags(currentObject);
+          return this.escapeHTMLTags(currentObject);
         return currentObject;
       }
     } catch (_) {
       throw new Error("i18n resource is not initialized");
     }
+  }
+
+  private interpolate(string: string, values: ILooseObject, escapeHTML = true) {
+    return string.replace(/\[(.*?)\]/g, (match, key) => {
+      const value = values[key.trim()] || match;
+      return escapeHTML ? this.escapeHTMLTags(value) : value;
+    });
+  }
+
+  private escapeHTMLTags(string: string) {
+    return string.replace(/[&<>"'`=/]/g, (match) => {
+      const escapeMap: Record<string, string> = {
+        // eslint-disable-next-line quotes
+        '"': "&quot;",
+        "&": "&amp;",
+        "'": "&#39;",
+        "/": "&#47;",
+        "<": "&lt;",
+        "=": "&#61;",
+        ">": "&gt;",
+        "`": "&#96;",
+      };
+      return escapeMap[match];
+    });
   }
 }