+++ packages/i18n-core/src/utils/appendQueryParams.util.ts
@@ -1,5 +1,0 @@
-export const appendQueryParams = (key: string, value: string) => {
-  const newURL = new URL(window.location.href);
-  newURL.searchParams.set(key, value);
-  window.history.replaceState({}, "", newURL.href);
-};